Update cal.c
In this improved version:

I added input validation for the operator and checked if it's one of the expected values. If not, it provides an error message and exits with an error code.

I added input validation for the numbers. It checks if both numbers were successfully scanned. If not, it provides an error message and exits with an error code.